Load with back emf is added
however there is a problem about dclink balancing in the three level NPC topology. The main component of the ripple is around 150hz, the switching frequency component has very low amplitude compared to this ripple. Moreover, the average voltage between the DCLINK capacitor vary a lot as the inverter goes to steady state , for 8888VA 0.9pf and 100uF capacitor(each), need to learn more about this